Removed license from source file → public domain
Previously, a comment at the top of the file listed userChrome.css as being released under a permissive modified ISC license.  This did not match the Unlicense commitment of the work to the public domain.  This modification places the Unlicense public domain commitment in the source file, to bring it in line with the repository's public domain commitment.